Moved attributes into '.Metadata' namespace
