+++ DSharpPlus.Commands/Processors/SlashCommands/Metadata/InteractionInstallTypeAttribute.cs
@@ -1,13 +1,13 @@
 using System;
 using System.Collections.Generic;
 
-namespace DSharpPlus.Commands.Processors.SlashCommands.Attributes;
+namespace DSharpPlus.Commands.Processors.SlashCommands.Metadata;
 
 /// <summary>
 /// Specifies the installation context for a command or module.
 /// </summary>
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = false)]
-public class InteractionInstallTypeAttribute(params ApplicationIntegrationType[]  installTypes) : Attribute
+public class InteractionInstallTypeAttribute(params ApplicationIntegrationType[] installTypes) : Attribute
 {
     /// <summary>
     /// The contexts the command is allowed to be installed to.
